+++ world-map-generator/src/main/java/net/world/map/structure/config/WorldConfig.java
@@ -1,6 +1,6 @@
 package net.world.map.structure.config;
 
 public class WorldConfig {
-    public static final int MAX_HEIGHT = 319;
-    public static final int MIN_HEIGHT = -64;
+    public static final short MAX_HEIGHT = 319;
+    public static final short MIN_HEIGHT = -64;
 }
