+++ world-map-generator/src/main/java/net/world/map/structure/model/UnderwaterBlock.java
@@ -3,7 +3,7 @@ package net.world.map.structure.model;
 public class UnderwaterBlock extends Block {
     public int depth;
 
-    public UnderwaterBlock(int x, int y, int height, int depth, String blockType) {
+    public UnderwaterBlock(int x, int y, short height, int depth, String blockType) {
         super(x, y, height, blockType);
         this.depth = depth;
     }
